sql/schemachanger/scexec/backfiller: extract some logic from scdeps
As a rule, we shouldn't be putting much business logic in scdeps. This change
creates a new package scexec/backfiller which implements the BackfillerTracker
and PeriodicProgressFlusher.

Release note: None
