+++ pkg/sql/BUILD.bazel
@@ -379,6 +379,7 @@ go_library(
         "//pkg/sql/schemachanger/scdeps",
         "//pkg/sql/schemachanger/scerrors",
         "//pkg/sql/schemachanger/scexec",
+        "//pkg/sql/schemachanger/scexec/backfiller",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan",
