+++ pkg/sql/schemachanger/scdeps/run_deps.go
@@ -22,6 +22,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descs"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scexec"
+	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scexec/backfiller"
 	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scrun"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
@@ -37,7 +38,7 @@ func NewJobRunDependencies(
 	internalExecutor sqlutil.InternalExecutor,
 	backfiller scexec.Backfiller,
 	merger scexec.Merger,
-	rangeCounter RangeCounter,
+	rangeCounter backfiller.RangeCounter,
 	eventLoggerFactory EventLoggerFactory,
 	jobRegistry *jobs.Registry,
 	job *jobs.Job,
@@ -82,7 +83,7 @@ type jobExecutionDeps struct {
 	backfiller            scexec.Backfiller
 	merger                scexec.Merger
 	commentUpdaterFactory scexec.DescriptorMetadataUpdaterFactory
-	rangeCounter          RangeCounter
+	rangeCounter          backfiller.RangeCounter
 	jobRegistry           *jobs.Registry
 	job                   *jobs.Job
 	kvTrace               bool
@@ -125,17 +126,14 @@ func (d *jobExecutionDeps) WithTxnInJob(ctx context.Context, fn scrun.JobTxnFunc
 			},
 			backfiller: d.backfiller,
 			merger:     d.merger,
-			backfillTracker: newBackfillerTracker(
+			backfillerTracker: backfiller.NewTracker(
 				d.codec,
-				newBackfillerTrackerConfig(ctx, d.codec, d.rangeCounter, d.job),
-				convertFromJobBackfillProgress(
-					d.codec, pl.GetNewSchemaChange().BackfillProgress,
-				),
-				convertFromJobMergeProgress(
-					d.codec, pl.GetNewSchemaChange().MergeProgress,
-				),
+				d.rangeCounter,
+				d.job,
+				pl.GetNewSchemaChange().BackfillProgress,
+				pl.GetNewSchemaChange().MergeProgress,
 			),
-			periodicProgressFlusher: newPeriodicProgressFlusherForIndexBackfill(d.settings),
+			periodicProgressFlusher: backfiller.NewPeriodicProgressFlusherForIndexBackfill(d.settings),
 			statements:              d.statements,
 			user:                    pl.UsernameProto.Decode(),
 			clock:                   NewConstantClock(timeutil.FromUnixMicros(pl.StartedMicros)),
