+++ pkg/sql/schemachanger/scexec/backfiller/periodic_progress_flusher_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package scdeps
+package backfiller
 
 import (
 	"context"
