+++ pkg/sql/schemachanger/scjob/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//pkg/sql",
         "//pkg/sql/schemachanger/scdeps",
         "//pkg/sql/schemachanger/scexec",
+        "//pkg/sql/schemachanger/scexec/backfiller",
         "//pkg/sql/schemachanger/scrun",
     ],
 )
